Move authorization to /utils
